requirements.txt
discord.py>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiohttp>=3.8.0
pytz>=2023.3